// src/blueprint/commands/command-types.ts
export enum CommandType {
  DRAW_PLOT = "DRAW_PLOT",
  DRAW_ROOM = "DRAW_ROOM",
  DRAW_DOOR = "DRAW_DOOR",
  DRAW_WINDOW = "DRAW_WINDOW",
  DRAW_TEXT = "DRAW_TEXT",
  DRAW_DIMENSION = "DRAW_DIMENSION",
  DRAW_STAIRCASE = "DRAW_STAIRCASE",
  DRAW_NORTH_ARROW = "DRAW_NORTH_ARROW",
  DRAW_SCALE_INDICATOR = "DRAW_SCALE_INDICATOR",
  DRAW_AREA_SUMMARY = "DRAW_AREA_SUMMARY",
}